requirements.txt
fastapi
google-cloud-firestore
jinja2
uvicorn[standard]